+++ src/main/java/entities/Person.java
@@ -4,7 +4,7 @@
  * a person not in a Group.
  */
 
-package Entities;
+package entities;
 
 public class Person {
     protected String name;
@@ -41,8 +41,8 @@ public class Person {
         return balance;
     }
 
-    public void setBalance(double balance) {
-        this.balance = balance;
+    public void updateBalance(double amount) {
+        this.balance += amount;
     }
 
     public String getEmail() {