+++ src/main/java/entities/User.java
@@ -3,7 +3,7 @@
  * and TODO: Insert functions
  */
 
-package Entities;
+package entities;
 
 import java.util.*;
 
@@ -11,6 +11,7 @@ import java.util.*;
 public class User extends Person implements AccountFeatures {
     private final String UUID;
     private List<Group> groups;
+    public List<String> expenses; // TODO make private
 
     /**
      * Construct User, giving them the given name, balance, and email.
@@ -21,6 +22,7 @@ public class User extends Person implements AccountFeatures {
     public User(String name, double balance, String email){
         super(name, balance, email);
         this.UUID = generateUUID();
+        this.expenses = new ArrayList<String>();
     }
 
     public User(double balance, String email){
@@ -37,7 +39,7 @@ public class User extends Person implements AccountFeatures {
         /*
         TODO: Implement this method
          */
-        return "Update generateUUID() in";
+        return "Update generateUUID() in User class";
     }
 
     public String getUUID(){