+++ src/test/entities/UserTest.java
@@ -1,10 +1,10 @@
-package java.Entities;
+package entities;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.*;
-import Entities.*;
+import entities.*;
 
 public class UserTest {